Delete GUI module and related functions
